requirements.txt
tqdm
argparse
matplotlib>2.0.0
numpy>=1.0.0
pandas
torchvision
tensorboard > 2.4.0
torch-tb-profiler
ipykernel >= 6.0.0
jupyter
dill
seaborn
sklearn
plotly
imbalanced-learn